% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotbyleader.r
\name{boxplotbyleader}
\alias{boxplotbyleader}
\title{Making boxplot. you need a dataset made by valueswithleader()}
\usage{
boxplotbyleader(dataSource = "")
}
\arguments{
\item{No}{parameter}
}
\value{
titleAbb+Boxplot
}
\description{
Making boxplot. you need a dataset made by valueswithleader()
}

