% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getindex.r
\name{getindex}
\alias{getindex}
\title{Get timeseries data from fred etc.}
\usage{
getindex(no = 1)
}
\arguments{
\item{no=1}{NIKKEI225 , no=2 DEXJPUS , no=3 JPNASSETS}
}
\value{
preData you can make your data as preData
}
\description{
Get timeseries data from fred etc.
}

