% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgb.r
\name{jgb}
\alias{jgb}
\title{Get Japanese Government Bond Interest Rates}
\usage{
jgb(year = 10)
}
\arguments{
\item{year(year=99:get}{all data.default year=10)}
}
\value{
jgbData preData
}
\description{
Get Japanese Government Bond Interest Rates
}

