% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldpvsdpj.r
\encoding{UTF-8}
\name{ldpvsdpj}
\alias{ldpvsdpj}
\title{Compare performance between Liberal Democratic Party of Japan and Democratic Party of Japan}
\usage{
ldpvsdpj()
}
\arguments{
\item{No}{parameter}
}
\value{
dpj ldp compareData timeSeries
}
\description{
Compare performance between Liberal Democratic Party of Japan and Democratic Party of Japan
}

