% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBaseData.r
\name{loadBaseData}
\alias{loadBaseData}
\title{Building data frame of leaders(onload function)}
\usage{
loadBaseData()
}
\arguments{
\item{No}{parameter}
}
\value{
pmoj goboj indexes indextitle titleAbb dataSource borderDate
}
\description{
Building data frame of leaders(onload function)
}

