% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHtml.r
\encoding{UTF-8}
\name{outputHtml}
\alias{outputHtml}
\title{Output data frame as html table}
\usage{
outputHtml(dataframe = compareData)
}
\arguments{
\item{No}{parameter}
}
\value{
htmldata
}
\description{
Output data frame as html table
}

