% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceAsLeader.r
\docType{package}
\name{performanceAsLeader}
\alias{performanceAsLeader}
\alias{performanceAsLeader-package}
\title{performanceAsLeader}
\description{
JGB interest rate - http://www.mof.go.jp/english/jgbs/reference/interest_rate/index.htm
NIKKEI225 DEXJPUS - https://research.stlouisfed.org/fred2/
}

