% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancesummary.r
\name{performancesummary}
\alias{performancesummary}
\title{Building data frame for performance summary by leaders}
\usage{
performancesummary()
}
\arguments{
\item{No}{parameter}
}
\value{
titleAbb+Summary
}
\description{
Building data frame for performance summary by leaders
}

