% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueswithleader.r
\name{valueswithleader}
\alias{valueswithleader}
\title{Building timeseries data with leaders}
\usage{
valueswithleader(target = 2, dd = 9999)
}
\arguments{
\item{target=target}{column , dd=day(99:end of month,999:middle of month,9999:as is,other:specific day}
}
\value{
dataset
}
\description{
Building timeseries data with leaders
}

